Remove test that skips work in scenario as scenario rarely occurs in the wild
The scenario is roughly that two dependencies of a single artifact have the same
groupId+artifactId+version. This does occur for some gwt projects where the source
dependency and the binary dependency are present at the same time. However in this
case the source artifact is already present locally so the guard against searching
for source does not save much overhead in the grand scheme of things.